[sdk]: fold the declaration codec into phantom aggregation on polkadot-util primitives
