+++ sdk/packages/sdk/src/protocols/intents/index.ts
@@ -26,7 +26,7 @@ export {
 	orderCommitment,
 } from "./utils"
 export { CryptoUtils, SELECT_SOLVER_TYPEHASH, PACKED_USEROP_TYPEHASH, DOMAIN_TYPEHASH } from "./CryptoUtils"
-export { encodeAcceptedSourceChains, decodeAcceptedSourceChains } from "./phantom-source-declaration"
+export { encodeAcceptedSourceChains, decodeAcceptedSourceChains } from "./phantom-aggregation"
 export {
 	DEFAULT_GRAFFITI,
 	ERC7821_BATCH_MODE,
